fixed warnings

+++ src/HighScore/HighScore.cpp
@@ -13,7 +13,7 @@ void HighScore::init() {
 }
 
 void HighScore::readHighScoresFromMemory() {
-    for (uint8_t i = 0; i < 5; ++i) {
+    for (uint16_t i = 0; i < 5; ++i) {
         uint16_t currentRead = eeprom_read_word((uint16_t*)i);
         if (currentRead != UINT16_T_MAX)
             highScores[i] = currentRead;
@@ -21,7 +21,7 @@ void HighScore::readHighScoresFromMemory() {
 }
 
 void HighScore::resetHighScores() {
-    for (uint8_t i = 0; i < 5; ++i) {
+    for (uint16_t i = 0; i < 5; ++i) {
         eeprom_write_word((uint16_t*)i, 0);
     }
     readHighScoresFromMemory();
@@ -48,7 +48,7 @@ uint8_t HighScore::getPlaceOnLeaderBoard() {
 }
 
 void HighScore::addHighScore(uint16_t score, uint8_t index) {
-    for (uint8_t i = 5; i >= index; --i) {
+    for (uint16_t i = 5; i >= index; --i) {
         highScores[i] = highScores[i-1];
     }
     highScores[index] = score;
